refactor: diffview_comments core is separated into standalone diffview-independent modules
